Move `RequestBindingGenerator` to `codegen-client`
