+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/smithy/generators/http/RestRequestSpecGenerator.kt
@@ -16,7 +16,7 @@ import software.amazon.smithy.rust.codegen.core.smithy.protocols.HttpBindingReso
 /**
  * [RestRequestSpecGenerator] generates a restJson1 or restXml specific `RequestSpec`. Both protocols are routed the same.
  *
- * This class has to live in the `codegen` subproject instead of in the `codegen-server` subproject because it is used
+ * This class has to live in the `codegen-core` subproject instead of in the `codegen-server` subproject because it is used
  * by the implementations of the `serverRouterRequestSpec` of the [Protocol] interface, which is used by both subprojects
  * (even though only the `codegen-server` subproject calls `serverRouterRequestSpec`).
  */
